requirements.txt
easydict==1.13
langchain==0.3.7
langchain-openai==0.2.6
lxml==5.3.0
mwparserfromhell==0.6.6
neo4j==5.26.0
openai==1.54.1
pandas==2.2.3
python-dotenv==1.0.1
tiktoken==0.8.0
torch==2.2.1
